requirements.txt
dlib==19.24.2
joblib==1.3.2
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4
threadpoolctl==3.2.0
types-Pillow==10.1.0.2
